feat: добавить класс popup и удалить файл modal
